Load chain info at app start
